Task 3 v1.2 redo basic controllers to rest api
