+++ src/main/java/ru/itmentor/spring/boot_security/demo/controller/UsersController.java
@@ -2,34 +2,28 @@ package ru.itmentor.spring.boot_security.demo.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.Authentication;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import ru.itmentor.spring.boot_security.demo.auth.UserDetailsImpl;
+import ru.itmentor.spring.boot_security.demo.dto.UserResponseDto;
 import ru.itmentor.spring.boot_security.demo.service.UserService;
+import ru.itmentor.spring.boot_security.demo.util.UserDtoConverter;
 
-@Controller
+@RestController
 @RequestMapping("/user")
 public class UsersController {
 
-    private UserService userService;
+    private final UserService userService;
+    private final UserDtoConverter userDtoConverter;
 
     @Autowired
-    public UsersController(UserService userService) {
+    public UsersController(UserService userService, UserDtoConverter userDtoConverter) {
         this.userService = userService;
+        this.userDtoConverter = userDtoConverter;
     }
 
-    @GetMapping()
-    public String getUserPage(Authentication authentication) {
-        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
-        return "redirect:user/"+userDetails.getUser().getId();
-    }
 
     @GetMapping("/{id}")
     @PreAuthorize("hasRole('ADMIN') or (hasRole('USER') and #id == principal.getUser().getId())")
-    public String getUser(@PathVariable int id, Model model) {
-        model.addAttribute("user", userService.getUser(id));
-        return "users/cabinet";
+    public UserResponseDto getUser(@PathVariable("id") int id) {
+        return userDtoConverter.toDto(userService.getUser(id));
     }
 }
