--- src/main/resources/templates/admin/edit.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Edit user</title>
-    <link rel="icon" href="data:,">
-</head>
-<body style="margin: 25px 50px 25px 50px">
-    <form th:method="PATCH" th:action="@{/admin/{id}(id=${user.getId()})}" th:object="${user}">
-        <label for="name">Enter Name: </label>
-        <input type="text" th:field="*{name}" id="name">
-        <div style="color:red" th:if="${#fields.hasErrors('name')}" th:errors="*{name}"></div>
-        <br><br>
-        <label for="age">Enter Age: </label>
-        <input type="text" th:field="*{age}" id="age">
-        <div style="color:red" th:if="${#fields.hasErrors('age')}" th:errors="*{age}"></div>
-        <br><br>
-        <label for="email">Enter Email: </label>
-        <input type="text" th:field="*{email}" id="email">
-        <div style="color:red" th:if="${#fields.hasErrors('email')}" th:errors="*{email}"></div>
-        <br><br>
-        <div th:if="${isCurUser}" th:each="role: ${user.getRoles()}">
-            <div th:text="${role.getRole()}"></div>
-        </div>
-        <div th:unless="${isCurUser}">
-            <label for="roles">Select Roles: </label>
-            <select th:field="*{roles}" multiple="multiple" id="roles">
-                <option th:each="role : ${roles}"
-                        th:value="${role.getId()}"
-                        th:text="${role.getRole()}"
-                        th:selected="${#lists.contains(user.getRoles(), role)}">
-                </option>
-            </select>
-        </div>
-        <div style="color:red" th:if="${#fields.hasErrors('roles')}" th:errors="*{roles}"></div>
-        <br><br>
-        <input type="submit" value="Edit">
-    </form>
-    <br><br>
-    <a href="/admin">Back to users...</a>
-</body>
-</html>
