--- src/main/resources/templates/admin/new.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>New user</title>
-    <link rel="icon" href="data:,">
-</head>
-<body style="margin: 25px 50px 25px 50px">
-    <form th:method="POST" th:action="@{/admin}" th:object="${user}">
-        <label for="name">Enter Name: </label>
-        <input type="text" th:field="*{name}" id="name">
-        <div style="color:red" th:if="${#fields.hasErrors('name')}" th:errors="*{name}"></div>
-        <br><br>
-        <label for="age">Enter Age: </label>
-        <input type="text" th:field="*{age}" id="age">
-        <div style="color:red" th:if="${#fields.hasErrors('age')}" th:errors="*{age}"></div>
-        <br><br>
-        <label for="email">Enter Email: </label>
-        <input type="text" th:field="*{email}" id="email">
-        <div style="color:red" th:if="${#fields.hasErrors('email')}" th:errors="*{email}"></div>
-        <br><br>
-        <label for="login">Enter Login: </label>
-        <input type="text" th:field="*{login}" id="login">
-        <div style="color:red" th:if="${#fields.hasErrors('login')}" th:errors="*{login}"></div>
-        <br><br>
-        <label for="password">Enter Password: </label>
-        <input type="text" th:field="*{password}" id="password">
-        <div style="color:red" th:if="${#fields.hasErrors('password')}" th:errors="*{password}"></div>
-        <br><br>
-        <label for="roles">Select Roles: </label>
-        <select th:field="*{roles}" multiple="multiple" id="roles">
-            <option th:each="role : ${roles}"
-                    th:value="${role.getId()}"
-                    th:text="${role.getRole()}"
-                    th:selected="${#lists.contains(user.getRoles(), role)}">
-            </option>
-        </select>
-        <div style="color:red" th:if="${#fields.hasErrors('roles')}" th:errors="*{roles}"></div>
-        <br><br>
-        <input type="submit" value="Add">
-    </form>
-    <br><br>
-    <a href="/admin">Back to users...</a>
-</body>
-</html>
