--- src/main/resources/templates/admin/user.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>User</title>
-    <link rel="icon" href="data:,">
-</head>
-<body style="margin: 25px 50px 25px 50px">
-    <div style="color: red" th:if="${user == null}" th:text="'User not found'"></div>
-    <div th:if="${user != null}">
-        <div>User Info</div>
-        <br><br>
-        <div th:text="'Name: '+${user.getName()}"></div>
-        <div th:text="'Age: '+${user.getAge()}"></div>
-        <div th:text="'Email: '+${user.getEmail()}"></div>
-        <p>Roles:</p>
-        <div th:each="role: ${user.getRoles()}">
-            <div th:text="${role.getRole()}"></div>
-        </div>
-        <br><br>
-        <form th:method="GET" th:action="@{/admin/{id}/edit(id=${user.getId()})}">
-            <input type="submit" value="Edit User">
-        </form>
-        <br><br>
-        <form th:method="DELETE" th:action="@{/admin/{id}(id=${user.getId()})}">
-            <input type="submit" value="Delete User" th:disabled="${isCurUser}">
-        </form>
-    </div>
-    <br><br>
-    <a href="/admin">Back to users...</a>
-    <br>
-    <br>
-    <br>
-    <form th:method="POST" th:action="@{/logout}">
-        <input type="submit" value="Logout">
-    </form>
-</body>
-</html>
