--- src/main/resources/templates/auth/login.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Login</title>
-</head>
-<body style="margin: 25px 50px 25px 50px">
-<form name="login" method="post" action="/process_login">
-    <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}"/>
-
-    <label for="username">Enter login: </label>
-    <input type="text" name="username" id="username">
-    <br>
-    <label for="password">Enter password: </label>
-    <input type="password" name="password" id="password">
-    <br>
-    <input type="submit" value="Login">
-
-    <div th:if="${param.error}" style="color: red">
-        Incorrect login or password
-    </div>
-</form>
-</body>
-</html>
