--- src/main/resources/templates/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="https://www.thymeleaf.org" xmlns:sec="https://www.thymeleaf.org/thymeleaf-extras-springsecurity3" lang="en">
-<head>
-    <title>Spring Security Example</title>
-</head>
-<body style="margin: 25px 50px 25px 50px">
-    <div th:if="${isAuthenticated}">
-        <h1 th:text="${username}+', Welcome to TEN APP!'"></h1>
-        <br>
-        <br>
-        <a th:text="Start" th:href="@{${nextPageUrl}}"></a>
-    </div>
-    <div th:unless="${isAuthenticated}">
-        <h1 th:text="'Hello, Welcome to TEN APP!'"></h1>
-        <br>
-        <br>
-        <a href="/auth/login">Log In</a>
-    </div>
-</body>
-</html>
